Revert "[pydrake] Move geometry files into subdir (#18684)"
This reverts commit 3aa7c8f345c31654b452b283f6c289f340ddee21.
