+++ bindings/pydrake/_geometry_extra.py
