--- bindings/pydrake/geometry/BUILD.bazel
@@ -1,164 +1,0 @@
-# -*- python -*-
-
-load("@drake//tools/install:install.bzl", "install")
-load(
-    "@drake//tools/skylark:drake_py.bzl",
-    "drake_py_unittest",
-)
-load(
-    "@drake//tools/skylark:pybind.bzl",
-    "add_pybind_coverage_data",
-    "drake_pybind_library",
-    "get_drake_py_installs",
-    "get_pybind_package_info",
-)
-load("//bindings/pydrake:pydrake.bzl", "add_lint_tests_pydrake")
-
-package(default_visibility = [
-    "//bindings:__subpackages__",
-])
-
-# This determines how `PYTHONPATH` is configured, and how to install the
-# bindings.
-PACKAGE_INFO = get_pybind_package_info(base_package = "//bindings")
-
-drake_pybind_library(
-    name = "geometry",
-    cc_deps = [
-        "//bindings/pydrake:documentation_pybind",
-        "//bindings/pydrake/common:default_scalars_pybind",
-        "//bindings/pydrake/common:deprecation_pybind",
-        "//bindings/pydrake/common:identifier_pybind",
-        "//bindings/pydrake/common:monostate_pybind",
-        "//bindings/pydrake/common:serialize_pybind",
-        "//bindings/pydrake/common:type_pack",
-        "//bindings/pydrake/common:type_safe_index_pybind",
-        "//bindings/pydrake/common:value_pybind",
-    ],
-    cc_so_name = "__init__",
-    cc_srcs = [
-        "geometry_py.cc",
-        "geometry_py.h",
-        "geometry_py_common.cc",
-        "geometry_py_hydro.cc",
-        "geometry_py_optimization.cc",
-        "geometry_py_render.cc",
-        "geometry_py_scene_graph.cc",
-        "geometry_py_visualizers.cc",
-    ],
-    package_info = PACKAGE_INFO,
-    py_data = [
-        "//setup:deepnote",
-    ],
-    py_deps = [
-        "//bindings/pydrake:math_py",
-        "//bindings/pydrake:module_py",
-        "//bindings/pydrake/solvers",
-        "//bindings/pydrake/systems:framework_py",
-        "//bindings/pydrake/systems:lcm_py",
-    ],
-    py_srcs = ["_geometry_extra.py"],
-)
-
-install(
-    name = "install",
-    targets = [":geometry"],
-    py_dest = PACKAGE_INFO.py_dest,
-    deps = get_drake_py_installs([":geometry"]),
-)
-
-drake_py_unittest(
-    name = "frame_id_test",
-    deps = [
-        ":geometry",
-    ],
-)
-
-drake_py_unittest(
-    name = "common_test",
-    deps = [
-        ":geometry",
-        "//bindings/pydrake/common",
-        "//bindings/pydrake/common/test_utilities",
-        "//bindings/pydrake/multibody:parsing_py",
-        "//bindings/pydrake/multibody:plant_py",
-    ],
-)
-
-drake_py_unittest(
-    name = "hydro_test",
-    data = [
-        "//geometry:test_obj_files",
-    ],
-    deps = [
-        ":geometry",
-        "//bindings/pydrake/common/test_utilities",
-    ],
-)
-
-drake_py_unittest(
-    name = "optimization_test",
-    deps = [
-        ":geometry",
-        "//bindings/pydrake:math_py",
-        "//bindings/pydrake/common/test_utilities:deprecation_py",
-        "//bindings/pydrake/common/test_utilities:pickle_compare_py",
-        "//bindings/pydrake/multibody:inverse_kinematics_py",
-        "//bindings/pydrake/multibody:parsing_py",
-        "//bindings/pydrake/multibody:plant_py",
-        "//bindings/pydrake/solvers",
-    ],
-)
-
-drake_py_unittest(
-    name = "render_test",
-    deps = [
-        ":geometry",
-        "//bindings/pydrake/common/test_utilities",
-        "//bindings/pydrake/systems:sensors_py",
-    ],
-)
-
-drake_py_unittest(
-    name = "render_engine_gl_test",
-    deps = [
-        ":geometry",
-        "//bindings/pydrake/common/test_utilities",
-    ],
-)
-
-drake_py_unittest(
-    name = "render_engine_subclass_test",
-    flaky = True,
-    deps = [
-        ":geometry",
-        "//bindings/pydrake/common/test_utilities",
-        "//bindings/pydrake/systems:sensors_py",
-    ],
-)
-
-drake_py_unittest(
-    name = "scene_graph_test",
-    deps = [
-        ":geometry",
-        "//bindings/pydrake/common/test_utilities",
-        "//bindings/pydrake/systems:sensors_py",
-    ],
-)
-
-drake_py_unittest(
-    name = "visualizers_test",
-    deps = [
-        ":geometry",
-        "//bindings/pydrake:lcm_py",
-        "//bindings/pydrake:perception_py",
-        "//bindings/pydrake/common/test_utilities",
-        "//bindings/pydrake/multibody:plant_py",
-        "//bindings/pydrake/systems:analysis_py",
-        "//bindings/pydrake/systems:lcm_py",
-    ],
-)
-
-add_pybind_coverage_data()
-
-add_lint_tests_pydrake()
