+++ bindings/pydrake/geometry_py.h
