+++ bindings/pydrake/geometry_py_common.cc
