+++ bindings/pydrake/geometry_py_hydro.cc
