+++ bindings/pydrake/geometry_py_render.cc
