+++ bindings/pydrake/geometry_py_scene_graph.cc
