+++ bindings/pydrake/geometry_py_visualizers.cc
