+++ bindings/pydrake/test/frame_id_test.py
