+++ bindings/pydrake/test/geometry_common_test.py
