+++ bindings/pydrake/test/geometry_hydro_test.py
