+++ bindings/pydrake/test/geometry_optimization_test.py
