+++ bindings/pydrake/test/geometry_render_engine_gl_test.py
