+++ bindings/pydrake/test/geometry_render_engine_subclass_test.py
