+++ bindings/pydrake/test/geometry_render_test.py
