+++ bindings/pydrake/test/geometry_scene_graph_test.py
