+++ bindings/pydrake/test/geometry_visualizers_test.py
